# openapi/control.json
{
  "components": {
    "responses": {
      "BadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Well-formed, but rejected by a rule; `detail.errors` lists each problem."
      },
      "Conflict": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "The resource's current state forbids this \u2014 a name already taken, a job already running."
      },
      "Error": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Unexpected error. Every error this API returns carries this body."
      },
...
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "No such resource in this organization."
      },
...
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Missing credential, or one that is revoked or no longer a member of this organization."
      },
      "ValidationFailed": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "The request did not match this operation's schema; `detail.errors` lists each field."
      }
    },
    "schemas": {
...
      "CreateTokenRequest": {
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateTokenRequest",
        "type": "object"
      },
      "ErrorBody": {
        "description": "The structured half of our error contract: a human `message` plus a list\nof individual `errors` (used wherever we surface many problems at once,\nincluding tool publish, op-tree checks, and request validation). The global\nexception handlers normalize every API error to\n`{detail: {message, errors}}`, so SDKs can parse one shape consistently.",
        "properties": {
          "errors": {
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message",
          "errors"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "The body of every error response, at every status code.",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "McpTokenResponse": {
        "description": "The credential half of an MCP client configuration.\n\nNo API URL: one token reaches every region, so the base URL is whichever\nregion the caller wants to work in \u2014 a choice only the dashboard is holding.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "token",
          "created_at"
        ],
        "title": "McpTokenResponse",
        "type": "object"
      },
...
      "OkResponse": {
        "description": "The generic '{ok: true}' body returned by deletes / no-content mutations.",
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "OkResponse",
        "type": "object"
      },
...
      "PageTokenResponse": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TokenResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "has_more",
          "limit",
          "offset"
        ],
        "title": "Page[TokenResponse]",
        "type": "object"
      },
...
      "TokenCreatedResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "manual",
              "mcp"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          },
          "user_email": {
            "title": "User Email",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "user_id",
          "user_email",
          "created_at",
          "token"
        ],
        "title": "TokenCreatedResponse",
        "type": "object"
      },
      "TokenResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "manual",
              "mcp"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "user_email": {
            "title": "User Email",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "user_id",
          "user_email",
          "created_at"
        ],
        "title": "TokenResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "SessionCookie": {
        "in": "cookie",
        "name": "talqing_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "summary": "Internal: identity, organizations, access tokens, payments, and the region list. Not published \u2014 the tenant-facing API is regional.",
    "title": "Talqing control API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/mcp-token": {
      "get": {
        "description": "This user's MCP token, in plaintext.\n\nUnlike a token created by hand, this one can be read whenever it is needed:\nthe dashboard renders it into ready-to-paste MCP configuration for Claude\nCode, Codex and Grok. The value is re-derived from the token's row, not\nstored, and is the same string every time \u2014 so a configuration a user\nalready saved keeps working.\n\nIt is always the caller's own token, never anyone else's, and is created on\nthe spot for a user who somehow has none. Delete it from the tokens list to\nrevoke it; the next call issues a new one, which rotates the credential.\n\nNo API URL comes with it, and it is not an omission: one token reaches every\nregion, and only the dashboard knows which region the person is looking at.\nControl could only answer with its own address, which is the one host an MCP\nclient must never be pointed at.",
        "operationId": "tokens.mcp",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Mcp Token",
        "tags": [
          "tokens"
        ]
      }
    },
...
    "/v1/tokens": {
      "get": {
        "description": "List personal access tokens: your own, or every one in the organization\nif you are an admin.\n\nWho holds which credential is organization-admin information, so an EDITOR\nor VIEWER sees only what they created. Plaintext token values are never\nreturned. A token with `kind: mcp` was issued automatically for the\ndashboard's MCP setup; the rest were created by hand.",
        "operationId": "tokens.list",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "List Tokens",
        "tags": [
          "tokens"
        ]
      },
      "post": {
        "description": "Create a personal access token for driving this API without the dashboard.\n\nThe token authenticates as its creator and acts with that user's live role,\nso it can never do more than they can. It does not expire; deleting it is\nthe revocation. The plaintext value is returned once here and never again \u2014\ntreat it like a password and do not echo it back into a conversation.",
        "operationId": "tokens.create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenCreatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Create Token",
        "tags": [
          "tokens"
        ]
      }
    },
    "/v1/tokens/{token_id}": {
      "delete": {
        "description": "Revoke a token immediately and permanently.\n\nOwners can delete their own; organization admins can delete any in the\norganization. Deleting an `mcp` token rotates it: the owner's next dashboard\nvisit issues a replacement, and anything still configured with the old one\nstops working.",
        "operationId": "tokens.delete",
        "parameters": [
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Token Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Delete Token",
        "tags": [
          "tokens"
        ]
      }
    }
  },
...
}